change type of field

+++ conveyance_app/models.py
@@ -14,8 +14,8 @@ class ConveyanceModel(models.Model):
     transport_mode = models.TextField(null=True, blank=True) 
     transport_cost = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     journey_status = models.CharField(max_length=20)
-    time_duration = models.IntegerField(null=True, blank=True) # in miliseconds
-    distance = models.IntegerField(null=True, blank=True) # in meters
+    time_duration = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    distance = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True) 
     created_at = models.DateTimeField(default=timezone.now)
     updated_at = models.DateTimeField(auto_now=True)
